Split capabilities page components
